+++ prefect_snowflake/__init__.py
@@ -1,7 +1,6 @@
 from . import _version
 from prefect_snowflake.credentials import (  # noqa
     SnowflakeCredentials,
-    SnowflakeConnector,
 )
 
 __version__ = _version.get_versions()["version"]