+++ prefect_snowflake/credentials.py
@@ -1,10 +1,9 @@
 """Credentials class to authenticate Snowflake."""
 
-from typing import Dict, Optional
+from typing import Optional
 
 from prefect.blocks.core import Block
 from pydantic import Field, SecretBytes, SecretStr, root_validator
-from snowflake import connector
 
 
 class SnowflakeCredentials(Block):
@@ -12,19 +11,19 @@ class SnowflakeCredentials(Block):
     Block used to manage authentication with Snowflake.
 
     Args:
-        account: The snowflake account name.
-        user: The user name used to authenticate.
-        password: The password used to authenticate.
-        private_key: The PEM used to authenticate.
-        authenticator: The type of authenticator to use for initializing
+        account (str): The snowflake account name.
+        user (str): The user name used to authenticate.
+        password (SecretStr): The password used to authenticate.
+        private_key (SecretStr): The PEM used to authenticate.
+        authenticator (str): The type of authenticator to use for initializing
             connection (oauth, externalbrowser, etc); refer to
             [Snowflake documentation](https://docs.snowflake.com/en/user-guide/python-connector-api.html#connect)
             for details, and note that `externalbrowser` will only
             work in an environment where a browser is available.
-        token: The OAuth or JWT Token to provide when
+        token (SecretStr): The OAuth or JWT Token to provide when
             authenticator is set to OAuth.
-        role: The name of the default role to use.
-        autocommit: Whether to automatically commit.
+        role (str): The name of the default role to use.
+        autocommit (bool): Whether to automatically commit.
 
     Example:
         Load stored Snowflake credentials:
@@ -37,14 +36,33 @@ class SnowflakeCredentials(Block):
     _block_type_name = "Snowflake Credentials"
     _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/2DxzAeTM9eHLDcRQx1FR34/f858a501cdff918d398b39365ec2150f/snowflake.png?h=250"  # noqa
 
-    account: str
-    user: str
-    password: Optional[SecretStr] = None
-    private_key: Optional[SecretBytes] = None
-    authenticator: Optional[str] = None
-    token: Optional[SecretStr] = None
-    role: Optional[str] = None
-    autocommit: Optional[bool] = None
+    account: str = Field(..., description="The snowflake account name")
+    user: str = Field(..., description="The user name used to authenticate")
+    password: Optional[SecretStr] = Field(
+        default=None, description="The password used to authenticate"
+    )
+    private_key: Optional[SecretBytes] = Field(
+        default=None, description="The PEM used to authenticate"
+    )
+    authenticator: Optional[str] = Field(
+        default=None,
+        description=(
+            "The type of authenticator to use for initializing "
+            "connection (oauth, externalbrowser, etc)"
+        ),
+    )
+    token: Optional[SecretStr] = Field(
+        default=None,
+        description=(
+            "The OAuth or JWT Token to provide when " "authenticator is set to oauth"
+        ),
+    )
+    role: Optional[str] = Field(
+        default=None, description="The name of the default role to use"
+    )
+    autocommit: Optional[bool] = Field(
+        default=None, description="Whether to automatically commit"
+    )
 
     @root_validator(pre=True)
     def _validate_auth_kwargs(cls, values):
@@ -58,92 +76,3 @@ class SnowflakeCredentials(Block):
                 f"One of the authentication keys must be provided: {auth_str}\n"
             )
         return values
-
-
-class SnowflakeConnector(Block):
-
-    """
-    Block used to manage connections with Snowflake.
-
-    Args:
-        database: The name of the default database to use.
-        warehouse: The name of the default warehouse to use.
-        schema: The name of the default schema to use.
-        credentials: The credentials to authenticate with Snowflake.
-
-    Example:
-        Load stored Snowflake connector:
-        ```python
-        from prefect_snowflake import SnowflakeConnector
-        snowflake_connector_block = SnowflakeConnector.load("BLOCK_NAME")
-        ```
-    """
-
-    _block_type_name = "Snowflake Connector"
-    _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/2DxzAeTM9eHLDcRQx1FR34/f858a501cdff918d398b39365ec2150f/snowflake.png?h=250"  # noqa
-
-    database: str
-    warehouse: str
-    schema_: str = Field(alias="schema")
-    credentials: SnowflakeCredentials
-
-    def _get_connect_params(self) -> Dict[str, str]:
-        """
-        Creates a connect params mapping to pass into get_connection.
-        """
-        connect_params = {
-            "database": self.database,
-            "warehouse": self.warehouse,
-            "schema": self.schema_,
-            # required to track task's usage in the Snowflake Partner Network Portal
-            "application": "Prefect_Snowflake_Collection",
-            **self.credentials.dict(),
-        }
-
-        # filter out unset values
-        connect_params = {
-            param: value for param, value in connect_params.items() if value is not None
-        }
-
-        for param in ("password", "private_key", "token"):
-            if param in connect_params:
-                connect_params[param] = connect_params[param].get_secret_value()
-
-        return connect_params
-
-    def get_connection(self) -> connector.SnowflakeConnection:
-        """
-        Returns an authenticated connection that can be
-        used to query from Snowflake databases.
-
-        Args:
-            database: The name of the database to use; overrides
-                the class definition if provided.
-            warehouse: The name of the warehouse to use; overrides
-                the class definition if provided.
-
-        Returns:
-            The authenticated SnowflakeConnection.
-
-        Examples:
-            ```python
-            from prefect import flow
-            from prefect_snowflake import SnowflakeConnector
-
-            @flow
-            def snowflake_connector_flow():
-                snowflake_connector = SnowflakeConnector(
-                    account="account",
-                    user="user",
-                    password="password",
-                    database="database",
-                    warehouse="warehouse",
-                )
-                return snowflake_connector
-
-            snowflake_connector_flow()
-            ```
-        """
-        connect_params = self._get_connect_params()
-        connection = connector.connect(**connect_params)
-        return connection